# cuda : add batched cuBLAS GEMM for faster attention