resolve #3479

## Description

For baseline performance on `master`, check the info in #3726

We identified the KQ and KQV operations in the attention layer as bottleneck. To overcome this, we parallelize the GEMMs across the head dimension using [cublasGemmBatchedEx](https://docs.nvidia.com/cuda/cublas/#cublasgemmbatchedex).

- The overall text-generation speedup with more than 1 batch is significant - up to 2-3 times compared to `master`
- The prompt processing speed is also improved for all models
- ~The single-batch text generation speed is improved for large contexts~ (it's the same, but maybe can be improved for large contexts)

I did some perplexity tests as well and I think everything works correctly, although I tested mostly F16 models. I don't think the accuracy of quantum models would be affected by this change, but still - need to double check.

Please take a careful look and let me know if you observe improved performance and if the ppl is good

## Results

Below are some results for A100. Also did similar tests for V100 and the results are very similar.

### Batched decoding for 7B F16 models

```
LLAMA_CUBLAS=1 make -j && ./batched-bench /workspace/openllama-7b/ggml-model-f16.gguf 4608 1 99 0 512 128 1,2,3,4,5,6,7,8,16,32
```

- master

|    PP |     TG |    B |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |      T s |    S t/s |
|-------|--------|------|--------|----------|----------|----------|----------|----------|----------|
|   512 |    128 |    1 |    640 |    0.126 |  4063.46 |    1.973 |    64.88 |    2.099 |   304.91 |
|   512 |    128 |    2 |    768 |    0.130 |  3941.89 |    6.521 |    39.26 |    6.651 |   115.48 |
|   512 |    128 |    3 |    896 |    0.118 |  4327.61 |    6.988 |    54.95 |    7.106 |   126.09 |
|   512 |    128 |    4 |   1024 |    0.111 |  4597.99 |    6.209 |    82.47 |    6.320 |   162.03 |
|   512 |    128 |    5 |   1152 |    0.110 |  4664.21 |    7.266 |    88.09 |    7.375 |   156.20 |
|   512 |    128 |    6 |   1280 |    0.108 |  4745.09 |    7.300 |   105.20 |    7.408 |   172.78 |
|   512 |    128 |    7 |   1408 |    0.111 |  4632.48 |    7.369 |   121.60 |    7.479 |   188.26 |
|   512 |    128 |    8 |   1536 |    0.111 |  4603.20 |    7.200 |   142.22 |    7.312 |   210.08 |
|   512 |    128 |   16 |   2560 |    0.111 |  4602.00 |    7.168 |   285.71 |    7.279 |   351.68 |
|   512 |    128 |   32 |   4608 |    0.113 |  4521.81 |    7.693 |   532.46 |    7.806 |   590.32 |

- PR

|    PP |     TG |    B |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |      T s |    S t/s |
|-------|--------|------|--------|----------|----------|----------|----------|----------|----------|
|   512 |    128 |    1 |    640 |    0.091 |  5648.97 |    1.839 |    69.61 |    1.929 |   331.71 |
|   512 |    128 |    2 |    768 |    0.087 |  5876.55 |    2.023 |   126.57 |    2.110 |   364.02 |
|   512 |    128 |    3 |    896 |    0.090 |  5690.47 |    2.064 |   186.01 |    2.154 |   415.89 |
|   512 |    128 |    4 |   1024 |    0.090 |  5683.21 |    2.096 |   244.30 |    2.186 |   468.46 |
|   512 |    128 |    5 |   1152 |    0.090 |  5661.59 |    2.158 |   296.58 |    2.248 |   512.38 |
|   512 |    128 |    6 |   1280 |    0.090 |  5687.18 |    2.185 |   351.47 |    2.275 |   562.61 |
|   512 |    128 |    7 |   1408 |    0.090 |  5709.32 |    2.231 |   401.60 |    2.321 |   606.70 |
|   512 |    128 |    8 |   1536 |    0.091 |  5645.36 |    2.268 |   451.60 |    2.358 |   651.34 |
|   512 |    128 |   16 |   2560 |    0.091 |  5640.63 |    2.784 |   735.57 |    2.875 |   890.43 |
|   512 |    128 |   32 |   4608 |    0.092 |  5550.26 |    2.841 |  1441.86 |    2.933 |  1571.08 |


### Batched decoding for 1B F16 models

```
LLAMA_CUBLAS=1 make -j && ./batched-bench /workspace/tinyllama-1b/ggml-model-f16.gguf 4608 1 99 0 512 128 1,2,3,4,5,6,7,8,16,32
```

- master

|    PP |     TG |    B |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |      T s |    S t/s |
|-------|--------|------|--------|----------|----------|----------|----------|----------|----------|
|   512 |    128 |    1 |    640 |    0.076 |  6716.34 |    0.662 |   193.33 |    0.738 |   866.82 |
|   512 |    128 |    2 |    768 |    0.060 |  8520.84 |    3.218 |    79.56 |    3.278 |   234.29 |
|   512 |    128 |    3 |    896 |    0.064 |  7995.38 |    3.258 |   117.86 |    3.322 |   269.70 |
|   512 |    128 |    4 |   1024 |    0.054 |  9441.96 |    3.302 |   155.05 |    3.356 |   305.08 |
|   512 |    128 |    5 |   1152 |    0.064 |  7943.90 |    3.259 |   196.39 |    3.323 |   346.65 |
|   512 |    128 |    6 |   1280 |    0.053 |  9593.22 |    3.325 |   230.98 |    3.378 |   378.88 |
|   512 |    128 |    7 |   1408 |    0.066 |  7770.41 |    3.425 |   261.62 |    3.491 |   403.36 |
|   512 |    128 |    8 |   1536 |    0.054 |  9464.83 |    3.463 |   295.66 |    3.518 |   436.67 |
|   512 |    128 |   16 |   2560 |    0.050 | 10192.71 |    3.369 |   607.92 |    3.419 |   748.74 |
|   512 |    128 |   32 |   4608 |    0.056 |  9119.08 |    3.770 |  1086.37 |    3.826 |  1204.24 |

- PR 
 
|    PP |     TG |    B |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |      T s |    S t/s |
|-------|--------|------|--------|----------|----------|----------|----------|----------|----------|
|   512 |    128 |    1 |    640 |    0.041 | 12337.94 |    0.560 |   228.57 |    0.602 |  1063.99 |
|   512 |    128 |    2 |    768 |    0.035 | 14805.39 |    0.831 |   307.97 |    0.866 |   887.01 |
|   512 |    128 |    3 |    896 |    0.027 | 18887.41 |    0.918 |   418.19 |    0.945 |   947.79 |
|   512 |    128 |    4 |   1024 |    0.040 | 12897.05 |    0.976 |   524.85 |    1.015 |  1008.65 |
|   512 |    128 |    5 |   1152 |    0.071 |  7234.19 |    0.983 |   651.16 |    1.054 |  1093.36 |
|   512 |    128 |    6 |   1280 |    0.040 | 12742.66 |    0.984 |   780.35 |    1.024 |  1249.58 |
|   512 |    128 |    7 |   1408 |    0.032 | 15873.02 |    1.017 |   880.62 |    1.050 |  1341.30 |
|   512 |    128 |    8 |   1536 |    0.040 | 12690.23 |    1.037 |   987.82 |    1.077 |  1426.22 |
|   512 |    128 |   16 |   2560 |    0.032 | 15780.07 |    1.309 |  1565.10 |    1.341 |  1909.04 |
|   512 |    128 |   32 |   4608 |    0.041 | 12502.75 |    1.374 |  2980.22 |    1.415 |  3255.74 |


### `llama-bench` for 7B F16

```
make -j && ../scripts/run-all-perf.sh openllama-7b "f16" "-ngl 99 -t 4 -p 1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,32,64,128,256,512,1024,2048 -n 128"
```

- master

| model                          |       size |     params | backend    | ngl |    threads | test       |              t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ---------: | ---------- | ---------------: |
| llama 7B mostly F16            |  12.55 GiB |     6.74 B | CUDA       |  99 |          4 | pp 1       |    37.82 ± 11.16 |
| llama 7B mostly F16            |  12.55 GiB |     6.74 B | CUDA       |  99 |          4 | pp 2       |     50.23 ± 0.07 |
| llama 7B mostly F16            |  12.55 GiB |     6.74 B | CUDA       |  99 |          4 | pp 3       |     73.45 ± 0.23 |
| llama 7B mostly F16            |  12.55 GiB |     6.74 B | CUDA       |  99 |          4 | pp 4       |    109.46 ± 0.08 |
| llama 7B mostly F16            |  12.55 GiB |     6.74 B | CUDA       |  99 |          4 | pp 5       |    120.57 ± 0.22 |
| llama 7B mostly F16            |  12.55 GiB |     6.74 B | CUDA       |  99 |          4 | pp 6       |    146.11 ± 0.14 |
| llama 7B mostly F16            |  12.55 GiB |     6.74 B | CUDA       |  99 |          4 | pp 7       |    166.63 ± 0.10 |
| llama 7B mostly F16            |  12.55 GiB |     6.74 B | CUDA       |  99 |          4 | pp 8       |    214.78 ± 0.17 |
| llama 7B mostly F16            |  12.55 GiB |     6.74 B | CUDA       |  99 |          4 | pp 9       |    212.03 ± 2.23 |
| llama 7B mostly F16            |  12.55 GiB |     6.74 B | CUDA       |  99 |          4 | pp 10      |    237.94 ± 0.12 |
| llama 7B mostly F16            |  12.55 GiB |     6.74 B | CUDA       |  99 |          4 | pp 11      |    259.16 ± 0.09 |
| llama 7B mostly F16            |  12.55 GiB |     6.74 B | CUDA       |  99 |          4 | pp 12      |    316.69 ± 2.23 |
| llama 7B mostly F16            |  12.55 GiB |     6.74 B | CUDA       |  99 |          4 | pp 13      |    302.72 ± 0.68 |
| llama 7B mostly F16            |  12.55 GiB |     6.74 B | CUDA       |  99 |          4 | pp 14      |    328.28 ± 0.88 |
| llama 7B mostly F16            |  12.55 GiB |     6.74 B | CUDA       |  99 |          4 | pp 15      |    347.02 ± 1.11 |
| llama 7B mostly F16            |  12.55 GiB |     6.74 B | CUDA       |  99 |          4 | pp 16      |    415.30 ± 1.38 |
| llama 7B mostly F16            |  12.55 GiB |     6.74 B | CUDA       |  99 |          4 | pp 32      |   721.61 ± 47.33 |
| llama 7B mostly F16            |  12.55 GiB |     6.74 B | CUDA       |  99 |          4 | pp 64      |   1338.41 ± 4.46 |
| llama 7B mostly F16            |  12.55 GiB |     6.74 B | CUDA       |  99 |          4 | pp 128     | 2538.48 ± 153.70 |
| llama 7B mostly F16            |  12.55 GiB |     6.74 B | CUDA       |  99 |          4 | pp 256     | 3924.85 ± 204.85 |
| llama 7B mostly F16            |  12.55 GiB |     6.74 B | CUDA       |  99 |          4 | pp 512     | 5089.59 ± 146.64 |
| llama 7B mostly F16            |  12.55 GiB |     6.74 B | CUDA       |  99 |          4 | pp 1024    |  4665.57 ± 24.93 |
| llama 7B mostly F16            |  12.55 GiB |     6.74 B | CUDA       |  99 |          4 | pp 2048    |   3421.92 ± 6.96 |
| llama 7B mostly F16            |  12.55 GiB |     6.74 B | CUDA       |  99 |          4 | tg 128     |     76.16 ± 0.11 |

- PR

| model                          |       size |     params | backend    | ngl |    threads | test       |              t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ---------: | ---------- | ---------------: |
| llama 7B mostly F16            |  12.55 GiB |     6.74 B | CUDA       |  99 |          4 | pp 1       |    62.60 ± 23.28 |
| llama 7B mostly F16            |  12.55 GiB |     6.74 B | CUDA       |  99 |          4 | pp 2       |    130.64 ± 1.33 |
| llama 7B mostly F16            |  12.55 GiB |     6.74 B | CUDA       |  99 |          4 | pp 3       |    194.78 ± 2.66 |
| llama 7B mostly F16            |  12.55 GiB |     6.74 B | CUDA       |  99 |          4 | pp 4       |    259.73 ± 1.01 |
| llama 7B mostly F16            |  12.55 GiB |     6.74 B | CUDA       |  99 |          4 | pp 5       |    315.80 ± 1.48 |
| llama 7B mostly F16            |  12.55 GiB |     6.74 B | CUDA       |  99 |          4 | pp 6       |    380.62 ± 0.99 |
| llama 7B mostly F16            |  12.55 GiB |     6.74 B | CUDA       |  99 |          4 | pp 7       |   435.28 ± 11.00 |
| llama 7B mostly F16            |  12.55 GiB |     6.74 B | CUDA       |  99 |          4 | pp 8       |    498.21 ± 3.11 |
| llama 7B mostly F16            |  12.55 GiB |     6.74 B | CUDA       |  99 |          4 | pp 9       |    550.94 ± 1.53 |
| llama 7B mostly F16            |  12.55 GiB |     6.74 B | CUDA       |  99 |          4 | pp 10      |   605.28 ± 13.77 |
| llama 7B mostly F16            |  12.55 GiB |     6.74 B | CUDA       |  99 |          4 | pp 11      |    665.26 ± 2.55 |
| llama 7B mostly F16            |  12.55 GiB |     6.74 B | CUDA       |  99 |          4 | pp 12      |    716.74 ± 7.20 |
| llama 7B mostly F16            |  12.55 GiB |     6.74 B | CUDA       |  99 |          4 | pp 13      |    774.57 ± 7.29 |
| llama 7B mostly F16            |  12.55 GiB |     6.74 B | CUDA       |  99 |          4 | pp 14      |   815.00 ± 19.86 |
| llama 7B mostly F16            |  12.55 GiB |     6.74 B | CUDA       |  99 |          4 | pp 15      |    874.26 ± 7.58 |
| llama 7B mostly F16            |  12.55 GiB |     6.74 B | CUDA       |  99 |          4 | pp 16      |    930.06 ± 9.51 |
| llama 7B mostly F16            |  12.55 GiB |     6.74 B | CUDA       |  99 |          4 | pp 32      | 1585.78 ± 230.53 |
| llama 7B mostly F16            |  12.55 GiB |     6.74 B | CUDA       |  99 |          4 | pp 64      |  2809.06 ± 53.23 |
| llama 7B mostly F16            |  12.55 GiB |     6.74 B | CUDA       |  99 |          4 | pp 128     | 4990.19 ± 434.57 |
| llama 7B mostly F16            |  12.55 GiB |     6.74 B | CUDA       |  99 |          4 | pp 256     | 6431.97 ± 506.03 |
| llama 7B mostly F16            |  12.55 GiB |     6.74 B | CUDA       |  99 |          4 | pp 512     | 6777.28 ± 324.30 |
| llama 7B mostly F16            |  12.55 GiB |     6.74 B | CUDA       |  99 |          4 | pp 1024    |  6441.81 ± 60.10 |
| llama 7B mostly F16            |  12.55 GiB |     6.74 B | CUDA       |  99 |          4 | pp 2048    |  5581.89 ± 49.59 |
| llama 7B mostly F16            |  12.55 GiB |     6.74 B | CUDA       |  99 |          4 | tg 128     |     76.13 ± 0.16 |


### Serving multiple clients with parallel decoding and continuous batching

```bash
# 128 sequences, 8 parallel
make -j && ./bin/parallel -m ../models/openllama-7b/ggml-model-f16.gguf -t 1 -ngl 100 -c 4096 -b 512 -s 1 -np 8 -ns 128 -n 100 -cb

main: n_parallel = 8, n_sequences = 128, cont_batching = 1, system tokens = 293
External prompt file: used built-in defaults
Model and path used:  ../models/openllama-7b/ggml-model-f16.gguf

Total prompt tokens:   2011, speed: 111.82 t/s
Total gen tokens:      6448, speed: 358.54 t/s
Total speed (AVG):           speed: 470.36 t/s
Cache misses:             0

llama_print_timings:        load time =    2515.70 ms
llama_print_timings:      sample time =     952.12 ms /  6576 runs   (    0.14 ms per token,  6906.72 tokens per second)
llama_print_timings: prompt eval time =   15236.39 ms /  8721 tokens (    1.75 ms per token,   572.38 tokens per second)
llama_print_timings:        eval time =     456.73 ms /    31 runs   (   14.73 ms per token,    67.87 tokens per second)
llama_print_timings:       total time =   17984.59 ms


# 128 sequences, 16 parallel
make -j && ./bin/parallel -m ../models/openllama-7b/ggml-model-f16.gguf -t 1 -ngl 100 -c 4096 -b 512 -s 1 -np 16 -ns 128 -n 100 -cb

main: n_parallel = 16, n_sequences = 128, cont_batching = 1, system tokens = 293
External prompt file: used built-in defaults
Model and path used:  ../models/openllama-7b/ggml-model-f16.gguf

Total prompt tokens:   2011, speed: 148.40 t/s
Total gen tokens:      6883, speed: 507.91 t/s
Total speed (AVG):           speed: 656.30 t/s
Cache misses:             0


llama_print_timings:        load time =    2484.72 ms
llama_print_timings:      sample time =     988.93 ms /  7011 runs   (    0.14 ms per token,  7089.52 tokens per second)
llama_print_timings: prompt eval time =   10928.88 ms /  9172 tokens (    1.19 ms per token,   839.24 tokens per second)
llama_print_timings:        eval time =     230.13 ms /    15 runs   (   15.34 ms per token,    65.18 tokens per second)
llama_print_timings:       total time =   13552.41 ms


# 128 sequences, 32 parallel

main: n_parallel = 32, n_sequences = 128, cont_batching = 1, system tokens = 293
External prompt file: used built-in defaults
Model and path used:  ../models/openllama-7b/ggml-model-f16.gguf

Total prompt tokens:   2011, speed: 200.03 t/s
Total gen tokens:      6491, speed: 645.64 t/s
Total speed (AVG):           speed: 845.67 t/s
Cache misses:             0


llama_print_timings:        load time =    2435.73 ms
llama_print_timings:      sample time =     923.99 ms /  6619 runs   (    0.14 ms per token,  7163.53 tokens per second)
llama_print_timings: prompt eval time =    7695.20 ms /  8790 tokens (    0.88 ms per token,  1142.27 tokens per second)
llama_print_timings:        eval time =     105.69 ms /     5 runs   (   21.14 ms per token,    47.31 tokens per second)
llama_print_timings:       total time =   10054.29 ms
```

Clients | Master t/s | PR t/s
-- | -- | --
8 | 247.30 | 470.36
16 | 368.59 | 656.30
32 | 422.33 | 845.67

After update with `cublasGemmBatchedEx` and Codellama 7B F16:

Clients | PR t/s
-- | -- 
4 | 306.07
8 | 487.71
12 | 616.16
16 | 707.03
20 | 802.23
24 | 829.71
32 | 857.19


### Perplexity for F16 7B

```bash
make -j perplexity && ./bin/perplexity -m ../models/openllama-7b/ggml-model-f16.gguf -f ./wikitext-2-raw/wiki.test.raw  -ngl 100 -nommq
```

- master

```
llm_load_tensors: ggml ctx size =    0.10 MB
llm_load_tensors: using CUDA for GPU acceleration
llm_load_tensors: mem required  =  250.10 MB
llm_load_tensors: offloading 32 repeating layers to GPU
llm_load_tensors: offloading non-repeating layers to GPU
llm_load_tensors: offloaded 35/35 layers to GPU
llm_load_tensors: VRAM used: 12603.02 MB
...................................................................................................
llama_new_context_with_model: n_ctx      = 512
llama_new_context_with_model: freq_base  = 10000.0
llama_new_context_with_model: freq_scale = 1
llama_kv_cache_init: offloading v cache to GPU
llama_kv_cache_init: offloading k cache to GPU
llama_kv_cache_init: VRAM kv self = 256.00 MB
llama_new_context_with_model: kv self size  =  256.00 MB
llama_new_context_with_model: compute buffer total size = 76.63 MB
llama_new_context_with_model: VRAM scratch buffer: 70.50 MB
llama_new_context_with_model: total VRAM used: 12929.52 MB (model: 12603.02 MB, context: 326.50 MB)

system_info: n_threads = 96 / 192 | AVX = 1 | AVX2 = 1 | AVX512 = 0 | AVX512_VBMI = 0 | AVX512_VNNI = 0 | FMA = 1 | NEON = 0 | ARM_FMA = 0 | F16C = 1 | FP16_VA = 0 | WASM_SIMD = 0 | BLAS = 1 | SSE3 = 1 | SSSE3 = 1 | VSX = 0 |
perplexity: tokenizing the input ..
perplexity: tokenization took 858.533 ms
perplexity: calculating perplexity over 616 chunks, batch_size=512
perplexity: 0.15 seconds per pass - ETA 1.55 minutes
…