@slaren 2023-10-23
`cublasGemmStridedBatchedEx` may also work, and it wouldn't require copying an array of pointers.