@KerfuffleV2 2023-10-23
I ran the `batched-bench` tests on my RX6600 with a Q5_K Mistral model. The results look weird, but this pull does seem to increase performance a little bit:

<details>

<summary>Expand</summary>

## PR

|    PP |     TG |    B |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |      T s |    S t/s |
|-------|--------|------|--------|----------|----------|----------|----------|----------|----------|
|   512 |    128 |    1 |    640 |    1.023 |   500.72 |    4.011 |    31.91 |    5.034 |   127.14 |
|   512 |    128 |    2 |    768 |    1.025 |   499.73 |   41.000 |     6.24 |   42.024 |    18.28 |
|   512 |    128 |    3 |    896 |    1.029 |   497.74 |   40.861 |     9.40 |   41.889 |    21.39 |
|   512 |    128 |    4 |   1024 |    1.025 |   499.31 |   40.541 |    12.63 |   41.566 |    24.64 |
|   512 |    128 |    5 |   1152 |    1.026 |   498.87 |   40.685 |    15.73 |   41.712 |    27.62 |
|   512 |    128 |    6 |   1280 |    1.029 |   497.39 |   41.978 |    18.30 |   43.008 |    29.76 |
|   512 |    128 |    7 |   1408 |    1.031 |   496.63 |   40.992 |    21.86 |   42.023 |    33.51 |
|   512 |    128 |    8 |   1536 |    1.029 |   497.56 |   41.915 |    24.43 |   42.944 |    35.77 |
|   512 |    128 |   16 |   2560 |    1.031 |   496.44 |   43.351 |    47.24 |   44.383 |    57.68 |
|   512 |    128 |   32 |   4608 |    1.033 |   495.83 |   52.089 |    78.63 |   53.122 |    86.74 |


## Master

|    PP |     TG |    B |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |      T s |    S t/s |
|-------|--------|------|--------|----------|----------|----------|----------|----------|----------|
|   512 |    128 |    1 |    640 |    1.016 |   503.94 |    4.318 |    29.65 |    5.334 |   119.99 |
|   512 |    128 |    2 |    768 |    1.013 |   505.28 |   48.181 |     5.31 |   49.195 |    15.61 |
|   512 |    128 |    3 |    896 |    1.018 |   502.93 |   46.285 |     8.30 |   47.303 |    18.94 |
|   512 |    128 |    4 |   1024 |    1.017 |   503.32 |   46.901 |    10.92 |   47.918 |    21.37 |
|   512 |    128 |    5 |   1152 |    1.019 |   502.61 |   47.411 |    13.50 |   48.429 |    23.79 |
|   512 |    128 |    6 |   1280 |    1.020 |   501.80 |   47.845 |    16.05 |   48.866 |    26.19 |
|   512 |    128 |    7 |   1408 |    1.021 |   501.69 |   47.975 |    18.68 |   48.996 |    28.74 |
|   512 |    128 |    8 |   1536 |    1.046 |   489.63 |   48.433 |    21.14 |   49.479 |    31.04 |
|   512 |    128 |   16 |   2560 |    1.020 |   501.90 |   53.560 |    38.24 |   54.580 |    46.90 |
|   512 |    128 |   32 |   4608 |    1.025 |   499.35 |   64.507 |    63.50 |   65.532 |    70.32 |

</details>

I tried messing with the MMQX/Y NWARPS stuff for Q5_K and Q6_K (the quants the model actually used) but it didn't seem to make much of a difference.