@ggerganov 2023-10-23
The MMQ constants take effect only if you run with MMQ=1 (set argv[5] to 1)